Fixed some comments. Moving to other laptop.
